Updates for brailleback 0.91.3 beta1.
git-svn-id: http://eyes-free.googlecode.com/svn@807 584082c0-ab3a-11dd-9ddb-6f86aeb5eef6

+++ trunk/braille/brailleback/src/com/googlecode/eyesfree/brailleback/BrailleView.java
@@ -16,6 +16,8 @@
 
 package com.googlecode.eyesfree.brailleback;
 
+import com.googlecode.eyesfree.braille.display.BrailleDisplayProperties;
+
 import android.content.Context;
 import android.content.res.TypedArray;
 import android.graphics.Canvas;
@@ -25,8 +27,6 @@ import android.util.AttributeSet;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewConfiguration;
-
-import com.googlecode.eyesfree.braille.display.BrailleDisplayProperties;
 
 /**
  * View which displays Braille dot patterns and corresponding text.
@@ -290,7 +290,7 @@ public class BrailleView extends View {
     public boolean onTouchEvent(MotionEvent event) {
         switch (event.getActionMasked()) {
             case MotionEvent.ACTION_DOWN:
-                int cellIndex = (int) ((float) event.getX() / mOuterWidth);
+                int cellIndex = (int) (event.getX() / mOuterWidth);
                 if (0 <= cellIndex && cellIndex < mNumTextCells) {
                     mPressedCell = cellIndex;
                     invalidate();
